+++ jest.config.json
@@ -1,5 +1,11 @@
 {
+  "collectCoverage": true,
   "coverageDirectory": "<rootDir>/coverage",
+  "coverageReporters": [
+    "html",
+    "json",
+    "text"
+  ],
   "moduleFileExtensions": [
     "js",
     "json",
