+++ package.json
@@ -14,7 +14,7 @@
     "@types/cookie-parser": "1.4.2",
     "@types/express": "4.17.13",
     "@types/jest": "27.4.0",
-    "@types/node": "17.0.12",
+    "@types/node": "17.0.13",
     "@types/supertest": "2.0.11",
     "@typescript-eslint/eslint-plugin": "5.10.1",
     "@typescript-eslint/parser": "5.10.1",
@@ -52,6 +52,7 @@
     "build:es6": "tsc -p src/tsconfig-es6.json",
     "clean": "rm -r es es6 lib",
     "test": "jest",
+    "test:coverage": "jest --coverage",
     "test:watch": "jest --watch",
     "pretest": "eslint ./",
     "publish-please": "publish-please",
