+++ src/constants.ts
@@ -37,8 +37,7 @@ export enum HTTP_VERBS_ENUM {
   trace = 'TRACE',
 }
 
-export const DUPLICATED_CONTROLLER_NAME = (name: string): string =>
-  `Two controllers cannot have the same name: ${name}`;
+export const DUPLICATED_CONTROLLER_NAME = (name: string): string => `Two controllers cannot have the same name: ${name}`;
 
 export const NO_CONTROLLERS_FOUND = 'No controllers' +
   'have been found! Please ensure that you have register' +
