+++ test/constants.test.ts
@@ -1,0 +1,10 @@
+import { DUPLICATED_CONTROLLER_NAME } from '../src/constants';
+
+describe('Constants test', () => {
+
+  it('should return correct message', () => {
+    expect(DUPLICATED_CONTROLLER_NAME('test'))
+      .toBe('Two controllers cannot have the same name: test');
+  });
+
+});
