+++ test/server.test.ts
@@ -1,139 +1,823 @@
-import express, { Application, NextFunction, Request, Response, Router } from 'express';
-import { Container } from 'inversify';
-import { InversifyExpressServer } from '../src/server';
-import { controller } from '../src/decorators';
+import supertest from 'supertest';
+import { interfaces } from 'inversify';
+import { Request, Response, Router, NextFunction, RequestHandler, json, CookieOptions } from 'express';
+import cookieParser from 'cookie-parser';
+import { injectable, Container } from 'inversify';
+import { AuthProvider, InversifyExpressServer, Principal } from '../src';
+import { controller, httpMethod, all, httpGet, httpPost, httpPut, httpPatch, httpHead, httpDelete, request, response, requestParam, requestBody, queryParam, requestHeaders, cookies, next, principal } from '../src/decorators';
 import { cleanUpMetadata } from '../src/utils';
-import { RoutingConfig } from '../src/interfaces';
-import { HttpResponseMessage } from '../src/httpResponseMessage';
 
-interface ServerWithTypes {
-  _app: Application;
-  _router: Router;
-  _routingConfig: RoutingConfig;
-  handleHttpResponseMessage: (
-    message: HttpResponseMessage,
-    res: Response
-  ) => void;
-}
+describe('Integration Tests:', () => {
+  let server: InversifyExpressServer;
+  let container: interfaces.Container;
 
-
-describe('Unit Test: InversifyExpressServer', () => {
   beforeEach(done => {
     cleanUpMetadata();
+    container = new Container();
     done();
   });
 
-  it('should call the configFn before the errorConfigFn', done => {
-    const middleware = (
-      req: Request,
-      res: Response,
-      next: NextFunction,
-    ) => {
-      //
-    };
+  describe('Routing & Request Handling:', () => {
+    it('should work for async controller methods', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/') public getTest(req: Request, res: Response) {
+          return new Promise((resolve => {
+            setTimeout(resolve, 100, 'GET');
+          }));
+        }
+      }
 
-    const configFn = jest.fn((app: Application) => {
-      app.use(middleware);
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .expect(200, 'GET', done);
     });
 
-    const errorConfigFn = jest.fn((app: Application) => {
-      app.use(middleware);
+    it('should work for async controller methods that fails', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/') public getTest(req: Request, res: Response) {
+          return new Promise(((resolve, reject) => {
+            setTimeout(reject, 100, 'GET');
+          }));
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .expect(500, done);
     });
 
-    const container = new Container();
+    it('should work for methods which call nextFunc()', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getTest(req: Request, res: Response, nextFunc: NextFunction) {
+          nextFunc();
+        }
 
-    @controller('/')
-    class TestController { }
+        @httpGet('/') public getTest2(req: Request, res: Response) {
+          return 'GET';
+        }
+      }
 
-    const server = new InversifyExpressServer(container);
-
-    server.setConfig(configFn)
-      .setErrorConfig(errorConfigFn);
-
-    expect(configFn).not.toBeCalled();
-    expect(errorConfigFn).not.toBeCalled();
-
-    server.build();
-
-    expect(configFn).toHaveBeenCalledTimes(1);
-    expect(errorConfigFn).toHaveBeenCalledTimes(1);
-    done();
-  });
-
-  it('Should allow to pass a custom Router instance as config', () => {
-    const container = new Container();
-
-    const customRouter = Router({
-      caseSensitive: false,
-      mergeParams: false,
-      strict: false,
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .expect(200, 'GET', done);
     });
 
-    const serverWithDefaultRouter = new InversifyExpressServer(container);
-    const serverWithCustomRouter = new InversifyExpressServer(
-      container,
-      customRouter
-    );
+    it('should work for async methods which call nextFunc()', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getTest(req: Request, res: Response, nextFunc: NextFunction) {
+          return new Promise((resolve => {
+            setTimeout(() => {
+              nextFunc();
+              resolve(null);
+            }, 100, 'GET');
+          }));
+        }
 
-    expect((serverWithDefaultRouter as unknown as ServerWithTypes)
-      ._router === customRouter).toBe(false);
-    expect((serverWithCustomRouter as unknown as ServerWithTypes)
-      ._router === customRouter).toBe(true);
-  });
+        @httpGet('/') public getTest2(req: Request, res: Response) {
+          return 'GET';
+        }
+      }
 
-  it('Should allow to provide custom routing configuration', () => {
-    const container = new Container();
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .expect(200, 'GET', done);
+    });
 
-    const routingConfig = {
-      rootPath: '/such/root/path',
-    };
+    it('should work for async methods called by nextFunc()', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getTest(req: Request, res: Response, nextFunc: NextFunction) {
+          return nextFunc;
+        }
 
-    const serverWithDefaultConfig = new InversifyExpressServer(container);
-    const serverWithCustomConfig = new InversifyExpressServer(
-      container,
-      null,
-      routingConfig
-    );
+        @httpGet('/') public getTest2(req: Request, res: Response) {
+          return new Promise((resolve => {
+            setTimeout(resolve, 100, 'GET');
+          }));
+        }
+      }
 
-    expect((serverWithCustomConfig as unknown as ServerWithTypes)
-      ._routingConfig).toBe(routingConfig);
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .expect(200, 'GET', done);
+    });
 
-    expect((serverWithDefaultConfig as unknown as ServerWithTypes)
-      ._routingConfig).not.toEqual(
-        (serverWithCustomConfig as unknown as ServerWithTypes)._routingConfig,
+    it('should work for each shortcut decorator', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getTest(req: Request, res: Response) { res.send('GET'); }
+
+        @httpPost('/')
+        public postTest(req: Request, res: Response) { res.send('POST'); }
+
+        @httpPut('/')
+        public putTest(req: Request, res: Response) { res.send('PUT'); }
+
+        @httpPatch('/')
+        public patchTest(req: Request, res: Response) { res.send('PATCH'); }
+
+        @httpHead('/')
+        public headTest(req: Request, res: Response) { res.send('HEAD'); }
+
+        @httpDelete('/')
+        public deleteTest(req: Request, res: Response) { res.send('DELETE'); }
+      }
+
+      server = new InversifyExpressServer(container);
+      const agent = supertest(server.build());
+
+      const deleteFn = () => {
+        void agent.delete('/').expect(200, 'DELETE', done);
+      };
+      const head = () => {
+        void agent.head('/').expect(200, 'HEAD', deleteFn);
+      };
+      const patch = () => { void agent.patch('/').expect(200, 'PATCH', head); };
+      const put = () => { void agent.put('/').expect(200, 'PUT', patch); };
+      const post = () => { void agent.post('/').expect(200, 'POST', put); };
+      const get = () => { void agent.get('/').expect(200, 'GET', post); };
+
+      get();
+    });
+
+    it('should work for more obscure HTTP methods using the httpMethod decorator', done => {
+      @controller('/')
+      class TestController {
+        @httpMethod('propfind', '/')
+        public getTest(req: Request, res: Response) {
+          res.send('PROPFIND');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .propfind('/')
+        .expect(200, 'PROPFIND', done);
+    });
+
+    it('should use returned values as response', done => {
+      const result = { hello: 'world' };
+
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getTest(req: Request, res: Response) {
+          return result;
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .expect(200, JSON.stringify(result), done);
+    });
+
+    it('should use custom router passed from configuration', () => {
+      @controller('/CaseSensitive')
+      class TestController {
+        @httpGet('/Endpoint') public get() {
+          return 'Such Text';
+        }
+      }
+
+      const customRouter = Router({
+        caseSensitive: true,
+      });
+
+      server = new InversifyExpressServer(container, customRouter);
+      const app = server.build();
+
+      const expectedSuccess = supertest(app)
+        .get('/CaseSensitive/Endpoint')
+        .expect(200, 'Such Text');
+
+      const expectedNotFound1 = supertest(app)
+        .get('/casesensitive/endpoint')
+        .expect(404);
+
+      const expectedNotFound2 = supertest(app)
+        .get('/CaseSensitive/endpoint')
+        .expect(404);
+
+      return Promise.all([
+        expectedSuccess,
+        expectedNotFound1,
+        expectedNotFound2,
+      ]);
+    });
+
+    it('should use custom routing configuration', () => {
+      @controller('/ping')
+      class TestController {
+        @httpGet('/endpoint') public get() {
+          return 'pong';
+        }
+      }
+
+      server = new InversifyExpressServer(
+        container,
+        null,
+        { rootPath: '/api/v1' }
       );
+
+      return supertest(server.build())
+        .get('/api/v1/ping/endpoint')
+        .expect(200, 'pong');
+    });
+
+    it('should work for controller methods who\'s return value is falsey', done => {
+      @controller('/user')
+      class TestController {
+        @httpDelete('/') public async delete(): Promise<void> {
+          //
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .delete('/user')
+        .expect(204, '', done);
+    });
   });
 
-  it('Should allow to provide a custom express application', () => {
-    const container = new Container();
-    const app = express();
-    const serverWithDefaultApp = new InversifyExpressServer(container);
-    const serverWithCustomApp = new InversifyExpressServer(
-      container,
-      null,
-      null,
-      app
-    );
-
-    expect((serverWithCustomApp as unknown as ServerWithTypes)._app).toBe(app);
-    expect((serverWithDefaultApp as unknown as ServerWithTypes)._app).not
-      .toEqual((serverWithCustomApp as unknown as ServerWithTypes)._app);
-  });
-
-  it('Should handle a HttpResponseMessage that has no content', () => {
-    const container = new Container();
-    const server = new InversifyExpressServer(container);
-
-    const httpResponseMessageWithoutContent = new HttpResponseMessage(404);
-    const mockResponse: Partial<Response> = {
-      sendStatus: jest.fn(),
+  describe('Middleware:', () => {
+    let result: string;
+    type Middleware = {
+      a: (req: Request, res: Response, nextFunc: NextFunction) => void;
+      b: (req: Request, res: Response, nextFunc: NextFunction) => void;
+      c: (req: Request, res: Response, nextFunc: NextFunction) => void;
+    };
+    const middleware: Middleware = {
+      a(req: Request, res: Response, nextFunc: NextFunction) {
+        result += 'a';
+        nextFunc();
+      },
+      b(req: Request, res: Response, nextFunc: NextFunction) {
+        result += 'b';
+        nextFunc();
+      },
+      c(req: Request, res: Response, nextFunc: NextFunction) {
+        result += 'c';
+        nextFunc();
+      },
     };
 
-    (server as unknown as ServerWithTypes).handleHttpResponseMessage(
-      httpResponseMessageWithoutContent,
-      mockResponse as unknown as Response,
-    );
+    const spyA = jest.fn().mockImplementation(middleware.a);
+    const spyB = jest.fn().mockImplementation(middleware.b);
+    const spyC = jest.fn().mockImplementation(middleware.c);
 
-    expect(mockResponse.sendStatus).toHaveBeenCalledWith(404);
+    beforeEach(done => {
+      spyA.mockClear();
+      spyB.mockClear();
+      spyC.mockClear();
+      result = '';
+      done();
+    });
+
+    it('should call method-level middleware correctly (GET)', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/', spyA, spyB, spyC)
+        public getTest(req: Request, res: Response) {
+          res.send('GET');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      const agent = supertest(server.build());
+
+      void agent.get('/')
+        .expect(200, 'GET', () => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(spyC).toHaveBeenCalledTimes(1);
+          expect(result).toBe('abc');
+          done();
+        });
+    });
+
+    it('should call method-level middleware correctly (POST)', done => {
+      @controller('/')
+      class TestController {
+        @httpPost('/', spyA, spyB, spyC)
+        public postTest(req: Request, res: Response) {
+          res.send('POST');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      const agent = supertest(server.build());
+
+      void agent.post('/')
+        .expect(200, 'POST', () => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(spyC).toHaveBeenCalledTimes(1);
+          expect(result).toBe('abc');
+          done();
+        });
+    });
+
+    it('should call method-level middleware correctly (PUT)', done => {
+      @controller('/')
+      class TestController {
+        @httpPut('/', spyA, spyB, spyC)
+        public postTest(req: Request, res: Response) {
+          res.send('PUT');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      const agent = supertest(server.build());
+
+      void agent.put('/')
+        .expect(200, 'PUT', () => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(spyC).toHaveBeenCalledTimes(1);
+          expect(result).toBe('abc');
+          done();
+        });
+    });
+
+    it('should call method-level middleware correctly (PATCH)', done => {
+      @controller('/')
+      class TestController {
+        @httpPatch('/', spyA, spyB, spyC)
+        public postTest(req: Request, res: Response) {
+          res.send('PATCH');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      const agent = supertest(server.build());
+
+      void agent.patch('/')
+        .expect(200, 'PATCH', () => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(spyC).toHaveBeenCalledTimes(1);
+          expect(result).toBe('abc');
+          done();
+        });
+    });
+
+    it('should call method-level middleware correctly (HEAD)', done => {
+      @controller('/')
+      class TestController {
+        @httpHead('/', spyA, spyB, spyC)
+        public postTest(req: Request, res: Response) {
+          res.send('HEAD');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      const agent = supertest(server.build());
+
+      void agent.head('/')
+        .expect(200, 'HEAD', () => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(spyC).toHaveBeenCalledTimes(1);
+          expect(result).toBe('abc');
+          done();
+        });
+    });
+
+    it('should call method-level middleware correctly (DELETE)', done => {
+      @controller('/')
+      class TestController {
+        @httpDelete('/', spyA, spyB, spyC)
+        public postTest(req: Request, res: Response) {
+          res.send('DELETE');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      const agent = supertest(server.build());
+
+      void agent.delete('/')
+        .expect(200, 'DELETE', () => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(spyC).toHaveBeenCalledTimes(1);
+          expect(result).toBe('abc');
+          done();
+        });
+    });
+
+    it('should call method-level middleware correctly (ALL)', done => {
+      @controller('/')
+      class TestController {
+        @all('/', spyA, spyB, spyC)
+        public postTest(req: Request, res: Response) {
+          res.send('ALL');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      const agent = supertest(server.build());
+
+      void agent.get('/')
+        .expect(200, 'ALL', () => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(spyC).toHaveBeenCalledTimes(1);
+          expect(result).toBe('abc');
+          done();
+        });
+    });
+
+    it('should call controller-level middleware correctly', done => {
+      @controller('/', spyA, spyB, spyC)
+      class TestController {
+        @httpGet('/')
+        public getTest(req: Request, res: Response) {
+          res.send('GET');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .expect(200, 'GET', () => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(spyC).toHaveBeenCalledTimes(1);
+          expect(result).toBe('abc');
+          done();
+        });
+    });
+
+    it('should call server-level middleware correctly', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getTest(req: Request, res: Response) {
+          res.send('GET');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+
+      server.setConfig(app => {
+        app.use(spyA);
+        app.use(spyB);
+        app.use(spyC);
+      });
+
+      void supertest(server.build())
+        .get('/')
+        .expect(200, 'GET', () => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(spyC).toHaveBeenCalledTimes(1);
+          expect(result).toBe('abc');
+          done();
+        });
+    });
+
+    it('should call all middleware in correct order', done => {
+      @controller('/', spyB)
+      class TestController {
+        @httpGet('/', spyC)
+        public getTest(req: Request, res: Response) {
+          res.send('GET');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+
+      server.setConfig(app => {
+        app.use(spyA);
+      });
+
+      void supertest(server.build())
+        .get('/')
+        .expect(200, 'GET', () => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(spyC).toHaveBeenCalledTimes(1);
+          expect(result).toBe('abc');
+          done();
+        });
+    });
+
+    it('should resolve controller-level middleware', () => {
+      const symbolId = Symbol.for('spyA');
+      const strId = 'spyB';
+
+      @controller('/', symbolId, strId)
+      class TestController {
+        @httpGet('/')
+        public getTest(req: Request, res: Response) {
+          res.send('GET');
+        }
+      }
+
+      container.bind<RequestHandler>(symbolId).toConstantValue(spyA);
+      container.bind<RequestHandler>(strId).toConstantValue(spyB);
+
+      server = new InversifyExpressServer(container);
+
+      const agent = supertest(server.build());
+
+      return agent.get('/')
+        .expect(200, 'GET')
+        .then(() => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(result).toBe('ab');
+        });
+    });
+
+    it('should resolve method-level middleware', () => {
+      const symbolId = Symbol.for('spyA');
+      const strId = 'spyB';
+
+      @controller('/')
+      class TestController {
+        @httpGet('/', symbolId, strId)
+        public getTest(req: Request, res: Response) {
+          res.send('GET');
+        }
+      }
+
+      container.bind<RequestHandler>(symbolId).toConstantValue(spyA);
+      container.bind<RequestHandler>(strId).toConstantValue(spyB);
+
+      server = new InversifyExpressServer(container);
+
+      const agent = supertest(server.build());
+
+      return agent.get('/')
+        .expect(200, 'GET')
+        .then(() => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(result).toBe('ab');
+        });
+    });
+
+    it('should compose controller- and method-level middleware', () => {
+      const symbolId = Symbol.for('spyA');
+      const strId = 'spyB';
+
+      @controller('/', symbolId)
+      class TestController {
+        @httpGet('/', strId)
+        public getTest(req: Request, res: Response) { res.send('GET'); }
+      }
+
+      container.bind<RequestHandler>(symbolId).toConstantValue(spyA);
+      container.bind<RequestHandler>(strId).toConstantValue(spyB);
+
+      server = new InversifyExpressServer(container);
+
+      const agent = supertest(server.build());
+
+      return agent.get('/')
+        .expect(200, 'GET')
+        .then(() => {
+          expect(spyA).toHaveBeenCalledTimes(1);
+          expect(spyB).toHaveBeenCalledTimes(1);
+          expect(result).toBe('ab');
+        });
+    });
+  });
+
+  describe('Parameters:', () => {
+    it('should bind a method parameter to the url parameter of the web request', done => {
+      @controller('/')
+      class TestController {
+        @httpGet(':id')
+        public getTest(
+          @requestParam('id') id: string,
+          req: Request,
+          res: Response
+        ) {
+          return id;
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/foo')
+        .expect(200, 'foo', done);
+    });
+
+    it('should bind a method parameter to the request object', done => {
+      @controller('/')
+      class TestController {
+        @httpGet(':id')
+        public getTest(
+          @request() req: Request
+        ) {
+          return req.params['id'];
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/GET')
+        .expect(200, 'GET', done);
+    });
+
+    it('should bind a method parameter to the response object', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getTest(
+          @response() res: Response
+        ) {
+          return res.send('foo');
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .expect(200, 'foo', done);
+    });
+
+    it('should bind a method parameter to a query parameter', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getTest(
+          @queryParam('id') id: string
+        ) {
+          return id;
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .query('id=foo')
+        .expect(200, 'foo', done);
+    });
+
+    it('should bind a method parameter to the request body', done => {
+      @controller('/')
+      class TestController {
+        @httpPost('/') public getTest(@requestBody() reqBody: string) {
+          return reqBody;
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      const body = { foo: 'bar' };
+      server.setConfig(app => {
+        app.use(json());
+      });
+      void supertest(server.build())
+        .post('/')
+        .send(body)
+        .expect(200, body, done);
+    });
+
+    it('should bind a method parameter to the request headers', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getTest(
+          @requestHeaders('testhead') headers: Record<string, unknown>) {
+          return headers;
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .set('TestHead', 'foo')
+        .expect(200, 'foo', done);
+    });
+
+    it('should be case insensitive to request headers', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getTest(
+          @requestHeaders('TestHead') headers: Record<string, unknown>) {
+          return headers;
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .set('TestHead', 'foo')
+        .expect(200, 'foo', done);
+    });
+
+    it('should bind a method parameter to a cookie', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/') public getCookie(
+          @cookies('Cookie') cookie: CookieOptions,
+          req: Request,
+          res: Response
+        ) {
+          return cookie;
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      server.setConfig(app => {
+        app.use(cookieParser());
+      });
+      void supertest(server.build())
+        .get('/')
+        .set('Cookie', 'Cookie=hey')
+        .expect(200, 'hey', done);
+    });
+
+    it('should bind a method parameter to the next function', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/') public getTest(@next() nextFunc: NextFunction) {
+          return nextFunc();
+        }
+
+        @httpGet('/') public getResult() {
+          return 'foo';
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .expect(200, 'foo', done);
+    });
+
+    it('should bind a method parameter to a principal with null (empty) details when no AuthProvider is set.', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getPrincipalTest(
+          @principal() userPrincipal: Principal
+        ) {
+          return userPrincipal.details;
+        }
+      }
+
+      server = new InversifyExpressServer(container);
+      void supertest(server.build())
+        .get('/')
+        .expect(200, '', done);
+    });
+
+    it('should bind a method parameter to a principal with valid details when an AuthProvider is set.', done => {
+      @controller('/')
+      class TestController {
+        @httpGet('/')
+        public getPrincipalTest(
+          @principal() userPrincipal: Principal
+        ) {
+          return userPrincipal.details;
+        }
+      }
+
+      @injectable()
+      class CustomAuthProvider implements AuthProvider {
+        public async getUser(
+          req: Request,
+          res: Response,
+          nextFunc: NextFunction,
+        ): Promise<Principal> {
+          return Promise.resolve({
+            details: 'something',
+            isAuthenticated: () => Promise.resolve(true),
+            isInRole: () => Promise.resolve(true),
+            isResourceOwner: () => Promise.resolve(true)
+          } as Principal);
+        }
+      }
+
+      server = new InversifyExpressServer(
+        container,
+        null,
+        null,
+        null,
+        CustomAuthProvider
+      );
+      void supertest(server.build())
+        .get('/')
+        .expect(200, 'something', done);
+    });
   });
 });
